+++ thesis/src/main/java/pls/PlsMapper.java
@@ -76,7 +76,7 @@ public abstract class PlsMapper extends MapReduceBase implements Mapper<BytesWri
 		//write out metadata before solutions
 		int offset = solutions.length * 4 * 2 + 4;
 		if (solutions.length > 1) {
-			dos.writeInt(solutions[0].getCost());
+			dos.writeDouble(solutions[0].getCost());
 			for (int i = 0; i < solutions.length; i++) {
 				dos.writeInt(offset);
 				dos.writeInt(solutions[i].serializedSize());
